Rename `shared::event::trigger` module into `shared::event::remote_targets`
The module now contains only the `RemoteTargets` trait.
