+++ src/shared/event.rs
@@ -3,6 +3,6 @@ pub mod client_trigger;
 pub mod ctx;
 pub mod event_fns;
 pub mod remote_event_registry;
+pub mod remote_targets;
 pub mod server_event;
 pub mod server_trigger;
-pub mod trigger;
