+++ src/shared/event/server_trigger.rs
@@ -9,8 +9,8 @@ use super::{
     ctx::{ClientReceiveCtx, ServerSendCtx},
     event_fns::{EventDeserializeFn, EventFns, EventSerializeFn},
     remote_event_registry::RemoteEventRegistry,
+    remote_targets::RemoteTargets,
     server_event::{self, ServerEvent, ToClients},
-    trigger::RemoteTargets,
 };
 use crate::shared::{backend::replicon_channels::Channel, entity_serde, postcard_utils};
